requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-rest-knox==4.2.0
django-cors-headers==3.13.0